Moved Traversable.to(java.lang.Iterable) to io.vavr.Iterable
